Allow creation of ProviderRequest by changing it to an interface, and adding a public create method.
