+++ security/security/src/main/java/io/helidon/security/OutboundSecurityClientImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2023 Oracle and/or its affiliates.
+ * Copyright (c) 2018, 2026 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import io.helidon.security.spi.OutboundSecurityProvider;
 
 /**
  * Outbound security builder and executor.
- *
+ * <p>
  * See {@link #submit()}.
  */
 final class OutboundSecurityClientImpl implements SecurityClient<OutboundSecurityResponse> {
@@ -42,8 +42,8 @@ final class OutboundSecurityClientImpl implements SecurityClient<OutboundSecurit
         this.security = security;
         this.context = context;
         this.providerName = providerName;
-        this.providerRequest = new ProviderRequest(context,
-                                                   request.resources());
+        this.providerRequest = ProviderRequest.create(context,
+                                                      request.resources());
         this.outboundEnv = outboundEnvironment;
         this.outboundEpConfig = outboundEndpointConfig;
     }
