+++ security/security/src/main/java/io/helidon/security/ProviderRequestImpl.java
@@ -1,0 +1,251 @@
+/*
+ * Copyright (c) 2018, 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.security;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+
+import io.helidon.security.util.AbacSupport;
+
+class ProviderRequestImpl implements ProviderRequest {
+    private static final System.Logger LOGGER = System.getLogger(ProviderRequestImpl.class.getName());
+
+    private final Map<String, AbacSupport> contextRoot = new HashMap<>();
+    private final Optional<Subject> subject;
+    private final Optional<Subject> service;
+    private final SecurityEnvironment env;
+    private final Optional<ObjectWrapper> resource;
+    private final SecurityContext context;
+    private final EndpointConfig epConfig;
+
+    ProviderRequestImpl(SecurityContext context,
+                        Map<String, Supplier<Object>> resources) {
+        ObjectWrapper object = null;
+
+        for (Map.Entry<String, Supplier<Object>> entry : resources.entrySet()) {
+            ObjectWrapper wrapper = new ObjectWrapper(entry.getValue());
+            contextRoot.put(entry.getKey(), wrapper);
+            if ("object".equals(entry.getKey())) {
+                object = wrapper;
+            }
+        }
+
+        this.env = context.env();
+        this.epConfig = context.endpointConfig();
+        this.context = context;
+        this.resource = Optional.ofNullable(object);
+        this.subject = context.user();
+        this.service = context.service();
+
+        contextRoot.put("env", env);
+        subject.ifPresent(user -> contextRoot.put("subject", user));
+        service.ifPresent(svc -> contextRoot.put("service", svc));
+    }
+
+    /**
+     * Get a value of a property from an object.
+     * If object implements {@link io.helidon.security.util.AbacSupport} the value is obtained through {@link io.helidon.security.util.AbacSupport#abacAttribute(String)}, if not,
+     * the value is obtained by reflection from a public field or a public getter method.
+     * The method name may be (for attribute called for example "audit"):
+     * <ul>
+     * <li>audit</li>
+     * <li>getAudit</li>
+     * <li>isAudit</li>
+     * <li>shouldAudit</li>
+     * <li>hasAudit</li>
+     * </ul>
+     *
+     * @param object object to get attribute from
+     * @param key    key of the attribute
+     * @return value of the attribute if found
+     */
+    static Optional<Object> getValue(Object object, String key) {
+        // use getter, public field
+        // first check if a public field with the name exists
+        Class<?> aClass = object.getClass();
+        try {
+            Field field = aClass.getField(key);
+            // java9
+            //if (field.canAccess(null)) {
+            // java8
+            if (ReflectionUtil.canAccess(ProviderRequestImpl.class, field)) {
+                return Optional.ofNullable(field.get(object));
+            }
+        } catch (NoSuchFieldException e) {
+            // ignore, field is not present, we try accessor methods
+            LOGGER.log(System.Logger.Level.TRACE, () ->
+                "Field \"" + key + "\" + is not present in class: " + aClass.getName(), e);
+        } catch (IllegalAccessException e) {
+            // ignore, we check access first
+            LOGGER.log(System.Logger.Level.TRACE, () ->
+                "Failed to access field: \"" + key + "\" in class: " + aClass.getName(), e);
+        }
+
+        //now check accessor methods
+        String capName = capitalize(key);
+        return getMethod(aClass, "get" + capName)
+                .or(() -> getMethod(aClass, key))
+                .or(() -> getMethod(aClass, "is" + capName))
+                .or(() -> getMethod(aClass, "has" + capName))
+                .or(() -> getMethod(aClass, "should" + capName))
+                .map(method -> {
+                    try {
+                        return method.invoke(object);
+                    } catch (Exception e) {
+                        throw new SecurityException("Failed to invoke method \"" + method + "\" on class \"" + aClass
+                                .getName() + "\"", e);
+                    }
+                });
+    }
+
+    static String capitalize(String string) {
+        char c = string.charAt(0);
+        char upperCase = Character.toUpperCase(c);
+        return upperCase + string.substring(1);
+    }
+
+    static Optional<Method> getMethod(Class<?> aClass, String methodName) {
+        try {
+            Method method = aClass.getMethod(methodName);
+            // java9
+            //if (method.canAccess(null)) {
+            // java8 not good approach (don't have any other)
+            if (ReflectionUtil.canAccess(ProviderRequestImpl.class, method)) {
+                return Optional.of(method);
+            }
+            return Optional.empty();
+        } catch (NoSuchMethodException e) {
+            // method is not present
+            LOGGER.log(System.Logger.Level.TRACE,
+                    () -> "Method: \"" + methodName + "\" is not in class: " + aClass.getName(), e);
+            return Optional.empty();
+        }
+    }
+
+    @Override
+    public EndpointConfig endpointConfig() {
+        return epConfig;
+    }
+
+    @Override
+    public SecurityContext securityContext() {
+        return context;
+    }
+
+    @Override
+    public Optional<Subject> subject() {
+        return subject;
+    }
+
+    @Override
+    public Optional<Subject> service() {
+        return service;
+    }
+
+    @Override
+    public SecurityEnvironment env() {
+        return env;
+    }
+
+    @Override
+    public Optional<Object> getObject() {
+        return resource.map(ObjectWrapper::getValue);
+    }
+
+    @Override
+    public Object abacAttributeRaw(String key) {
+        return contextRoot.get(key);
+    }
+
+    @Override
+    public Collection<String> abacAttributeNames() {
+        return contextRoot.keySet();
+    }
+
+    private static final class ObjectWrapper implements AbacSupport {
+        private final Supplier<Object> valueSupplier;
+        private volatile Object value;
+        private volatile AbacSupport container;
+
+        private ObjectWrapper(Supplier<Object> value) {
+            this.valueSupplier = value;
+        }
+
+        @Override
+        public Object abacAttributeRaw(String key) {
+            checkValue();
+
+            if (null != container) {
+                return container.abacAttributeRaw(key);
+            }
+            return ProviderRequestImpl.getValue(value, key);
+        }
+
+        @Override
+        public Collection<String> abacAttributeNames() {
+            checkValue();
+
+            if (null != container) {
+                return container.abacAttributeNames();
+            }
+
+            throw new UnsupportedOperationException("Property names are not available for general Object types, such as: "
+                                                            + value.getClass());
+        }
+
+        Object getValue() {
+            checkValue();
+            return value;
+        }
+
+        private void checkValue() {
+            if (null == value) {
+                synchronized (valueSupplier) {
+                    if (null == value) {
+                        value = valueSupplier.get();
+                        if (value instanceof AbacSupport) {
+                            container = (AbacSupport) value;
+                        } else if (value instanceof Map) {
+                            final Map<?, ?> map = (Map<?, ?>) value;
+                            container = new AbacSupport() {
+                                @Override
+                                public Object abacAttributeRaw(String key) {
+                                    return map.get(key);
+                                }
+
+                                @Override
+                                public Collection<String> abacAttributeNames() {
+                                    return map.keySet()
+                                            .stream()
+                                            .map(String::valueOf)
+                                            .collect(Collectors.toSet());
+                                }
+                            };
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+}
